Update average_time_to_start field type from int to float in ServerFields class and corresponding alembic migration

+++ src/sc_crawler/alembic/versions/f0273cb124f6_v0_6_0_add_new_server_fields.py
@@ -475,7 +475,7 @@ def upgrade() -> None:
             batch_op.add_column(
                 sa.Column(
                     "average_time_to_start",
-                    sa.Integer(),
+                    sa.Float(),
                     nullable=True,
                     comment="Average time to start the server (seconds).",
                 ),
@@ -565,7 +565,7 @@ def upgrade() -> None:
             server_table_name,
             sa.Column(
                 "average_time_to_start",
-                sa.Integer(),
+                sa.Float(),
                 nullable=True,
                 comment="Average time to start the server (seconds).",
             ),
