+++ src/sc_crawler/table_bases.py
@@ -768,7 +768,7 @@ class ServerFields(
     ipv4: int = Field(
         default=0, description="Number of complimentary IPv4 address(es)."
     )
-    average_time_to_start: Optional[int] = Field(
+    average_time_to_start: Optional[float] = Field(
         default=None,
         description="Average time to start the server (seconds).",
     )
